+++ relay-server/src/services/store.rs
@@ -766,8 +766,6 @@ impl StoreService {
         let scoping = message.scoping();
         let received_at = message.received_at();
 
-        let relay_emits_accepted_outcome = false;
-
         let meta = SpanMeta {
             organization_id: scoping.organization_id,
             project_id: scoping.project_id,
@@ -776,7 +774,6 @@ impl StoreService {
             retention_days: message.retention_days,
             downsampled_retention_days: message.downsampled_retention_days,
             received: datetime_to_timestamp(received_at),
-            accepted_outcome_emitted: relay_emits_accepted_outcome,
             performance_issues_spans: message.performance_issues_spans,
         };
 
@@ -802,20 +799,6 @@ impl StoreService {
             counter(RelayCounters::SpanV2Produced) += 1,
             via = "processing"
         );
-
-        if relay_emits_accepted_outcome {
-            // XXX: Temporarily produce span outcomes. Keep in sync with either EAP
-            // or the segments consumer, depending on which will produce outcomes later.
-            self.outcome_aggregator.send(TrackOutcome {
-                category: DataCategory::SpanIndexed,
-                event_id: None,
-                outcome: Outcome::Accepted,
-                quantity: 1,
-                remote_addr: None,
-                scoping,
-                timestamp: received_at,
-            });
-        }
 
         Ok(())
     }
@@ -1692,8 +1675,6 @@ struct SpanMeta {
     retention_days: u16,
     /// Number of days until the downsampled version of this data should be deleted.
     downsampled_retention_days: u16,
-    /// Indicates whether Relay already emitted an accepted outcome or if EAP still needs to emit it.
-    accepted_outcome_emitted: bool,
     /// Whether the segment span should be used for issue detection instead of the transaction.
     #[serde(rename = "_performance_issues_spans", skip_serializing_if = "is_false")]
     performance_issues_spans: bool,